This commit adds e2e tests for go/v3-alpha and refactors test code such that it is easier to add other plugin tests going forward.
test/e2e/utils: add util to remove tool binaries from path

test/e2e/v3: refactor generation code so all plugins can be tested

test/e2e/v2: rename test files

*: clean up scripts
