+++ test_e2e_local.sh
@@ -50,4 +50,4 @@ fi
 
 # when changing these commands, make sure to keep in sync with ./test_e2e.sh
 go test ./test/e2e/v2 -v -ginkgo.v
-go test ./test/e2e/v3 -v -ginkgo.v
+go test ./test/e2e/v3 -v -ginkgo.v -timeout 15m
